index: add function dispatcher
