+++ index/lib.h
@@ -41,8 +41,9 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
-#include "mx.h"
+#include "mutt/lib.h"
 #include "core/lib.h"
+#include "mx.h"          // IWYU pragma: keep
 #include "shared_data.h" // IWYU pragma: keep
 
 struct Context;
